added admin component

+++ src/app/show-files/show-files.component.ts
@@ -2,6 +2,8 @@ import { Component, OnInit } from '@angular/core'
 import { Router, ActivatedRoute } from '@angular/router'
 import { FileUploadServiceService } from '../_services/file-upload-service.service'
 import { AuthenticationService } from '../_services/authentication.service'
+import { MatSnackBar } from '@angular/material/snack-bar'
+
 
 @Component({
   selector: 'app-show-files',
@@ -12,9 +14,14 @@ export class ShowFilesComponent implements OnInit {
   files: string[] = []
   math = Math
   selection: string[] = []
+  deleteAllFiles: string[] = []
+
+  
+
 
   constructor(
     private router: Router,
+    private _snackBar: MatSnackBar,
     private fileUploadService: FileUploadServiceService,
     private authenticationService: AuthenticationService,
   ) {}
@@ -23,7 +30,7 @@ export class ShowFilesComponent implements OnInit {
 
     this.show_data()
   }
-
+  
   show_data() {
     console.log(this.authenticationService.currentID())
     this.fileUploadService.download(this.authenticationService.currentID()).subscribe(data => {
@@ -35,6 +42,16 @@ export class ShowFilesComponent implements OnInit {
         }
       }
     })
+  }
+
+
+  selectAll()
+  {
+
+    for (let file in this.files)
+      {
+      this.deleteAllFiles.push(this.files[file]['id'])
+      }
   }
 
 
@@ -50,18 +67,46 @@ export class ShowFilesComponent implements OnInit {
   }
 
   deleteFile() {
-    console.log(this.selection)
-    this.fileUploadService.file_delete(this.selection).subscribe(data => {
-      for (const key in data) {
-        if (key === 'success') {
-          console.log(data.success)
-          this.selection = [];
-          this.ngOnInit();
-        } else {
-          console.log('error')
+    if (this.selection.length === 0) {
+      return this._snackBar.open('No file selected', 'ok', {
+        duration: 3000,
+      });
+    }
+    else {
+      this.fileUploadService.file_delete(this.selection).subscribe(data => {
+        for (const key in data) {
+          if (key === 'success') {
+            console.log(data.success)
+            this.selection = [];
+            this.ngOnInit();
+          } else {
+            console.log('error')
+          }
         }
-      }
-    })
+      })
+    }
     
+  }
+
+  deleteAll() {
+    this.selectAll()
+    if (this.deleteAllFiles.length === 0) {
+      return this._snackBar.open('No file selected', 'ok', {
+        duration: 3000,
+      });
+    }
+    else {
+      this.fileUploadService.file_delete(this.deleteAllFiles).subscribe(data => {
+        for (const key in data) {
+          if (key === 'success') {
+            this.selection = [];
+            this.deleteAllFiles = [];
+            this.ngOnInit();
+          } else {
+            console.log('error')
+          }
+        }
+      })
+    }
   }
 }
